rename did module to client
