+++ client/encryptor.py
