+++ docs/client.rst
@@ -1,0 +1,15 @@
+client package
+===========
+
+Submodules
+----------
+
+.. toctree::
+
+   client.did
+   client.encryptor
+   client.enums
+   client.keys
+   client.models
+   client.service
+   client.updater
