--- docs/did.rst
@@ -1,15 +1,0 @@
-did package
-===========
-
-Submodules
-----------
-
-.. toctree::
-
-   did.did
-   did.encryptor
-   did.enums
-   did.keys
-   did.models
-   did.service
-   did.updater
